requirements.txt
fastapi
uvicorn[standard]
langgraph>=0.2.0
langchain-google-vertexai>=2.0.0
langchain-core>=0.3.0
google-cloud-bigquery
google-cloud-storage>=2.10.0
db-dtypes
google-auth
pandas
matplotlib
plotly
python-dotenv
